drivers: dma: dma-axi-dmac-test: implement basic test
The test is based on the Xilinx AXI DMA test & Atmel base DMA test, but
it's simplified for the case that's being specifically tested.

It needs a special HDL, with an AXI FIFO stream.
FIFO stream details:
  https://www.xilinx.com/support/documentation/ip_documentation/axi_fifo_mm_s/v4_1/pg080-axi-fifo-mm-s.pdf

The over-all schematic is:

|----------|
|   DDR    |
------------
     ^
     |
|----------|
|   PS7    |<----------------|
------------              64 AXI MM
     | AXI MM                |
|-----------|             |------|
| FIFO MM2S |---32 AXIS-->| DMAC |
-------------             --------

The DMA core is placed to offset 0x43C20000
The FIFO core is placed to offset 0x43C00000

The test requires that an external entity configures the FIFO, then a sysfs
call be done: `echo 1 > /sys/module/dma_axi_dmac_test/parameters/start_run`

The DMAC test will run a single DMA transfer to read data from the FIFO.
For this test, the data should be a null terminated string, but this can be
extended.

A sample DT for this test, on a Zed-Board:
```
/dts-v1/;

/include/ "zynq-zed.dtsi"
/include/ "zynq-zed-adv7511.dtsi"

&fpga_axi {

        axi_dmac_dmatest_0: axi_dmac_dmatest@0 {
                compatible ="adi,axi-dmac-test-1.00.a";
                dmas = <&axi_dmac 0>;
                dma-names = "axidma0";
        };

        axi_dmac: dmac@43c20000 {
                compatible = "adi,axi-dmac-1.00.a";
                reg = <0x43c20000 0x10000>;
                #dma-cells = <1>;
                interrupts = <0 56 0>;
                clocks = <&clkc 16>;

                adi,channels {
                        #size-cells = <0>;
                        #address-cells = <1>;

                        dma-channel@0 {
                                reg = <0>;
                                adi,source-bus-width = <32>;
                                adi,source-bus-type = <AXI_DMAC_BUS_TYPE_AXI_STREAM>;
                                adi,destination-bus-width = <64>;
                                adi,destination-bus-type = <AXI_DMAC_BUS_TYPE_AXI_MM>;
                                adi,length-width = <24>;
                        };
                };
        };
};

```

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>