dma: axi-dmac: terminate early DMA transfers when they're partial
When we get a partial transfer, we should not submit any more segments to
the hardware, as the data it will get will be unused and will just do
un-necessary processing.

So, we just increase the counters for submitted & completed segments to the
last index for partial transfers and that should finish the transfer early.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>