spi-axi-engine: Calculate buffer dimension for xfer
After the modification to the word length the buffer for rx and tx needed
to be adapted to the new size. Before this update the size was fixed to 8
bits and that was stored inside memory.

Example: 20 bit resolution -> entire 20 bit value stored inside 32
bit register. Value is separated into 3 parts of 8 bits and stored in the
rx buffer. The same for tx part.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>